Crate stats_traits

Library for calculating statistics on collections of numbers.

Re-exports

pub use crate::types::Result;

Modules

types

Module with type aliases

Enums

StatsError

Error type for the crate

Traits

FrequencyStats

Similar to Stats, but for collections of tuples containing a frequency and a value.

Stats

A trait to be implemented for collection-like types that provides statistical methods. Requires that the type it is implemented on can be converted into an Iterator.