Trait stats_traits::FrequencyStats

pub trait FrequencyStats<T>: IntoIterator<Item = (usize, T)> + Clonewhere
    T: NumExt,{
    fn count(&self) -> usize { ... }
    fn non_zero_count(&self) -> Result<usize> { ... }
    fn non_zero_count_into_item(&self) -> Result<T> { ... }
    fn sum(&self) -> Result<T> { ... }
    fn mean(&self) -> Result<T> { ... }
    fn variance(&self) -> Result<T> { ... }
    fn std_dev(&self) -> Result<T>
    where
        T: ToPrimitive,
    { ... }
}

Similar to Stats, but for collections of tuples containing a frequency and a value.

Provided Methods

fn count(&self) -> usize

Returns the total number of values in the collection. This is the sum of all frequencies.

Examples

use stats_traits::FrequencyStats;
let vec: Vec<(usize, i32)> = vec![(1, 1), (2, 2)];
assert_eq!(vec.count(), 3);

fn non_zero_count(&self) -> Result<usize>

Count the items in the collection, returning Err(StatsError::EmptyCollection) if it’s empty

fn non_zero_count_into_item(&self) -> Result<T>

Count the items in the collection and convert the result to Self::Item. Return an error under the same conditions as crate::Stats::non_zero_count, or if the length could not be converted to Self::Item

fn sum(&self) -> Result<T>

Calculate the sum of all the items in the collection multiplied by their frequency.

Examples

use stats_traits::FrequencyStats;
let vec: Vec<(usize, i32)> = vec![(1, 1), (2, 2)];
assert_eq!(vec.sum(), Ok(5));

use stats_traits::FrequencyStats;
let vec: Vec<(usize, f64)> = vec![(1, 1.0), (2, 2.0)];
assert_eq!(vec.sum(), Ok(5.0));

Errors

Returns an error if the frequency could not be converted to Self::Item

fn mean(&self) -> Result<T>

Calculate the mean of the collection. This is the sum of all the items in the collection multiplied by their frequency, divided by the total number of items. This is the same as crate::Stats::mean, but for collections of tuples containing a frequency and a value.

fn variance(&self) -> Result<T>

Calculate the variance of the collection. See crate::Stats::variance or Wikipedia for more information.

fn std_dev(&self) -> Result<T>where
T: ToPrimitive,

Calculate the standard deviation of the collection.

See crate::Stats::std_dev