Trait stats_traits::Stats

pub trait Stats: IntoIterator + Clonewhere
    Self::Item: NumExt,{
    fn sum(&self) -> Self::Item { ... }
    fn count(&self) -> usize { ... }
    fn non_zero_count(&self) -> Result<usize> { ... }
    fn non_zero_count_into_item(&self) -> Result<Self::Item> { ... }
    fn mean(&self) -> Result<Self::Item> { ... }
    fn variance(&self) -> Result<Self::Item> { ... }
    fn std_dev(&self) -> Result<Self::Item>
    where
        Self::Item: ToPrimitive,
    { ... }
    fn min(&self) -> Result<Self::Item>
    where
        Self::Item: MinMax,
    { ... }
    fn max(&self) -> Result<Self::Item>
    where
        Self::Item: MinMax,
    { ... }
    fn range(&self) -> Result<Self::Item>
    where
        Self::Item: MinMax,
    { ... }
}

A trait to be implemented for collection-like types that provides statistical methods. Requires that the type it is implemented on can be converted into an Iterator.

Provided Methods

fn sum(&self) -> Self::Item

Calculate the sum of all the items in the collection

Example

use stats_traits::Stats;
assert_eq!(vec![1, 2, 3].sum(), 6);

use stats_traits::Stats;
assert_eq!(vec![1.0, 2.0, 3.0].sum(), 6.0);

fn count(&self) -> usize

Count the items in the collection

Example

use stats_traits::Stats;
assert_eq!(vec![1, 2, 3].count(), 3);

fn non_zero_count(&self) -> Result<usize>

Count the items in the collection, returning Err(StatsError::EmptyCollection) if it’s empty

fn non_zero_count_into_item(&self) -> Result<Self::Item>

Count the items in the collection and convert the result to Self::Item. Return an error under the same conditions as Stats::non_zero_count, or if the length could not be converted to Self::Item

fn mean(&self) -> Result<Self::Item>

Find the mean of the collection

Examples

use stats_traits::Stats;
assert_eq!(vec![1, 2, 3].mean(), Ok(2));

use stats_traits::Stats;
let v: Vec<f64> = vec![1.0, 2.0, 3.0];
assert_eq!(v.mean(), Ok(2.0));

Watch out for integer division!

use stats_traits::Stats;
assert_eq!(vec![1, 2, 3, 4].mean(), Ok(2));

Errors

Returns an error if the collection is empty (has a length of 0). Will also return an error if the length of the collection is too large to fit in Self::Item.

fn variance(&self) -> Result<Self::Item>

Find the variance of the collection. The variance is the sum of the squared differences between each item and the mean, divided by the number of items in the collection. The variance is a measure of how spread out the items are.

Examples

use stats_traits::Stats;
assert_eq!(vec![1.0, 2.0, 3.0].variance(), Ok(2.0 / 3.0));

Errors

Errors under the same conditions as Stats::mean and Stats::non_zero_count_into_item

Wikipedia

fn std_dev(&self) -> Result<Self::Item>where
Self::Item: ToPrimitive,

Find the standard deviation of the collection. The standard deviation is the square root of the variance. It is a measure of how spread out the items are.

Wikipedia

Examples

use stats_traits::Stats;
use approx::assert_relative_eq;
assert_relative_eq!(vec![1.0, 2.0, 3.0].std_dev().unwrap(), 2.0_f64.sqrt() / 3.0_f64.sqrt());

Errors

Returns an error under the same conditions as Stats::variance

fn min(&self) -> Result<Self::Item>where
Self::Item: MinMax,

Return the smallest item in the collection

fn max(&self) -> Result<Self::Item>where
Self::Item: MinMax,

Return the largest item in the collection

fn range(&self) -> Result<Self::Item>where
Self::Item: MinMax,

Return the range of the collection (the smallest subtracted from the largest)